requirements.txt
torch
transformers
faiss-cpu
fastapi
uvicorn
pillow
numpy
pydantic
sentence-transformers
python-dotenv